# docs/swagger.yaml
basePath: /v1
definitions:
  constant.Gender:
    enum:
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - Male
    - Female
  model.Person:
    properties:
      dates:
        type: integer
      gender:
        $ref: '#/definitions/constant.Gender'
      height:
        type: number
      name:
        type: string
    type: object
  web.APIError:
    properties:
      errorCode:
        type: integer
      errorMessage:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: API Support
    url: https://CoulsonChen.github.io/
  description: API documentation.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Swagger API
  version: "1.0"
paths:
  /Match/{person_id}/{person_to_match_id}:
    put:
      consumes:
      - application/json
      description: Match people
      parameters:
      - description: Person ID
        in: path
        name: person_id
        required: true
        type: integer
      - description: Person-to-match ID
        in: path
        name: person_to_match_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: We need ID!!
          schema:
            $ref: '#/definitions/web.APIError'
      summary: Match people
  /Match/NewPerson:
    post:
      consumes:
      - application/json
      description: get string by ID
      parameters:
      - description: Person
        in: body
        name: Name
        required: true
        schema:
          $ref: '#/definitions/model.Person'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: We need ID!!
          schema:
            $ref: '#/definitions/web.APIError'
        "404":
          description: Can not find ID
          schema:
            $ref: '#/definitions/web.APIError'
      summary: Add a new pet to the store
  /Match/QueryMatches/{person_id}:
    get:
      consumes:
      - application/json
      description: Query matches for person
      parameters:
      - description: Person ID
        in: path
        name: person_id
        required: true
        type: integer
      - description: Top N Matches
        in: query
        name: top_n
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: We need ID!!
          schema:
            $ref: '#/definitions/web.APIError'
      summary: Query matches for person
  /Match/Remove/{person_id}:
    delete:
      consumes:
      - application/json
      description: Remove person from sorted set
      parameters:
      - description: Person ID
        in: path
        name: person_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
        "400":
          description: We need ID!!
          schema:
            $ref: '#/definitions/web.APIError'
      summary: Remove person from sorted set
swagger: "2.0"